requirements.txt
azure-ai-agents==1.2.0b3
azure-ai-projects==1.1.0b3
azure-identity==1.24.0
azure-monitor-opentelemetry==1.8.0
azure-search-documents==11.5.3
chainlit==2.7.2
openai==1.107.1
opentelemetry-instrumentation-openai==0.47.0
python-dotenv==1.1.1
semantic-kernel==1.36.2